requirements.txt
fastapi==0.95.1
openai==0.27.5
pydantic==1.10.7
python-dotenv==1.0.0
twilio==8.1.0
uvicorn==0.21.1